# Yangseng_API/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/upload": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "example": "any"
                },
                "employee_name": {
                  "example": "any"
                },
                "f_image": {
                  "example": "any"
                },
                "b_image": {
                  "example": "any"
                },
                "l_image": {
                  "example": "any"
                },
                "r_image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "502": {
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/api/getUserImages/{user_name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user_name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/getDateImages/{user_name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "datetime",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/upimage": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "example": "any"
                },
                "employee_name": {
                  "example": "any"
                },
                "f_image": {
                  "example": "any"
                },
                "b_image": {
                  "example": "any"
                },
                "l_image": {
                  "example": "any"
                },
                "r_image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/buf": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "example": "any"
                },
                "employee_name": {
                  "example": "any"
                },
                "f_image": {
                  "example": "any"
                },
                "b_image": {
                  "example": "any"
                },
                "l_image": {
                  "example": "any"
                },
                "r_image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}